OPC DA: Upgrade to OPCClientToolKit with 64bit support
Depend on 32/64bit OPCClientToolKit:
https://github.com/kumajaya/OPC-Client-X64.git

This upgrade includes:
* Source code adjustments due to library changes
* Completely disconnect a client from the server when disconnecting
* Remove FB group and items on terminated
* Update installation instructions

+++ src/com/opc/opccomlayer.cpp
@@ -99,11 +99,11 @@ EComResponse COpcComLayer::processInterrupt(){
       switch (mConnectionState){
       case e_Connected: {
         CIEC_ANY **apoRDs = mFb->getRDs();
-        unsigned int nrRDs = mFb->getNumRD();
+        size_t nrRDs = mFb->getNumRD();
 
         TOpcProcessVarList::iterator itEnd = mFBOutputVars.end();
         TOpcProcessVarList::iterator it = mFBOutputVars.begin();
-        for(unsigned int i = 0; i < nrRDs && it != itEnd; i++, ++it){
+        for(size_t i = 0; i < nrRDs && it != itEnd; i++, ++it){
           setOutputValue(&apoRDs[i]->unwrap(), &(*it)->updateValue());
         }
